'将Integer类型的id修改为Long'

+++ src/main/java/com/jfeat/am/module/order/api/ExpressEndpoint.java
@@ -60,7 +60,7 @@ public class ExpressEndpoint {
 
     @PutMapping("/{id}")
     @ApiOperation(value = "修改 Express", response = Express.class)
-    public Tip updateExpress(@PathVariable Integer id, @RequestBody Express entity) {
+    public Tip updateExpress(@PathVariable Long id, @RequestBody Express entity) {
         entity.setId(id);
         return SuccessTip.create(expressService.updateMaster(entity));
     }
@@ -89,7 +89,7 @@ public class ExpressEndpoint {
                               @RequestParam(name = "pageNum", required = false, defaultValue = "1") Integer pageNum,
                               @RequestParam(name = "pageSize", required = false, defaultValue = "10") Integer pageSize,
                               @RequestParam(name = "search", required = false) String search,
-                              @RequestParam(name = "id", required = false) Integer id,
+                              @RequestParam(name = "id", required = false) Long id,
                               @RequestParam(name = "code", required = false) String code,
                               @RequestParam(name = "name", required = false) String name,
                               @RequestParam(name = "enabled", required = false) Integer enabled,
